Crud das outras classes, arrumar authentication

+++ src/main/java/br/com/dbserver/uhpoa/models/Triagem.java
@@ -11,14 +11,13 @@ import javax.validation.constraints.NotBlank;
 
 @Entity
 @Table(name="TRIAGEM")
-
 public class Triagem implements Serializable{
 	
 	private static final long serialVersionUID = 1L;
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private Long id;
 	
 	
 	private String sintomasRelatadosPeloCliente;
@@ -38,10 +37,12 @@ public class Triagem implements Serializable{
 	private boolean perdaDoPaladar;
 	
 	
-	public int getId() {
+	
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getSintomasRelatadosPeloCliente() {
